% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGraphs.R
\docType{data}
\name{annoKidney}
\alias{annoKidney}
\title{Kidney sample annotation}
\format{
A dataframe with 3 columns
}
\usage{
data(annoKidney)
}
\description{
Tumor type annotation of the 40 kidney metabolic graphs provided with the
package
}
\keyword{datasets}
