% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceComp.R
\name{extractDist4part}
\alias{extractDist4part}
\title{Construct the distance matrix from 4 parts obtained from parallel run}
\usage{
extractDist4part(distance4partList)
}
\arguments{
\item{distance4partList}{List containing 4 parts of the distance matrix}
}
\value{
Gram matrix with pairwise distances between graphs
}
\description{
Construct the distance matrix from 4 parts obtained from parallel run
}
