% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionComp.R
\name{getBins}
\alias{getBins}
\title{Bins for the node distance distribution (NDD)}
\usage{
getBins(gList)
}
\arguments{
\item{gList}{List of graphs in igraph format}
}
\value{
binList
}
\description{
getBins finds the maximum diameter from the given graph list and returns the
bins needed for the  getNodeDistanceDistr function
}
\examples{
\dontrun{
data(KidneyGraphs)
binsG <- getBins(KidneyGraphs)}
}
\author{
Ichcha Manipur & Mario R Guarracino
}
