% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceComp.R
\name{getGraphdistance4PartsParallel}
\alias{getGraphdistance4PartsParallel}
\title{Parallel run of the graph distances computation}
\usage{
getGraphdistance4PartsParallel(distribMatlist, returnnodeDist = FALSE)
}
\arguments{
\item{distribMatlist}{list of probability distribution matrices}

\item{returnnodeDist}{Return node-wise distances(default=FALSE), if TRUE returns
a list containing 2 elements: 1) nodewise distances and 2) graph distance
(Gram) matrix}
}
\value{
Gram matrix with pairwise distances between graphs
}
\description{
Parallel run of the graph distances computation
}
\examples{
\dontrun{
data("KidneyGraphs")
binsList <- getBins(KidneyGraphs)
library(future.apply)
library(future)
plan(multiprocess)
nddListPar <- future_lapply(KidneyGraphs, function(x) getNodeDistanceDistr(x, binsList))
nddDistanceMultiPar <- getGraphdistance4PartsParallel(nddListPar)}
}
\author{
Ichcha Manipur
}
