% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceComp.R
\name{getGraphlistdistance}
\alias{getGraphlistdistance}
\title{Find the pairwise Jensen Shannon divergence between distribution matrices of
a list of graphs}
\usage{
getGraphlistdistance(distribMatlist, writenodeDist = FALSE)
}
\arguments{
\item{distribMatlist}{list of probability distribution matrices}
}
\value{
Gram matrix with pairwise distances between graphs
}
\description{
Find the pairwise Jensen Shannon divergence between distribution matrices of
a list of graphs
}
\examples{
\dontrun{
data("KidneyGraphs")
binsList <- getBins(KidneyGraphs)
nddList <- lapply(KidneyGraphs, function(x) getNodeDistanceDistr(x, binsaList))
getGraphlistdistance(nddList)}
}
