% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceComp.R
\name{getGraphlistdistance4Parts}
\alias{getGraphlistdistance4Parts}
\title{getGraphlistdistance4Parts}
\usage{
getGraphlistdistance4Parts(distribMatlist, index = 1)
}
\arguments{
\item{distribMatlist}{list of probability distribution matrices}

\item{index}{possible values 1-4}
}
\value{
Gram matrix with Pairwise distances
}
\description{
used when there are numerous graphs, the graph samples can be
divided into four as specified by the index. Run in 4 separate R sessions.
Save the 4 matrices and reassemble the gram matrix
}
