% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionComp.R
\name{getNodeDistanceDistr}
\alias{getNodeDistanceDistr}
\title{Node Distance Distribution computation}
\usage{
getNodeDistanceDistr(g, binList = NA)
}
\arguments{
\item{g}{graph in igraph format}

\item{binList}{bins for the distance distribution}
}
\value{
Node distance distribution
}
\description{
Node Distance Distribution computation
}
\examples{
\dontrun{
g <- igraph::make_tree(10, mode = "out")
getNodeDistanceDistr(g)}
}
\author{
Ichcha Manipur & Mario R Guarracino
}
