% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubnetworks.R
\name{getSubnetworksEigenadjacent}
\alias{getSubnetworksEigenadjacent}
\title{Extract subnetworks using Eigen centrality scores of nodes}
\usage{
getSubnetworksEigenadjacent(g, eigenSortcutoff = 0.1, percentTotalnodes = NA)
}
\arguments{
\item{g}{graph in igraph format}

\item{eigenSortcutoff}{cutoff based on eigen centrality score (values >0 and
<1) - default = 0.1}

\item{percentTotalnodes}{cutoff based on percentage of top eigen central
ranked nodes to be selected}
}
\value{
subnetwork
}
\description{
A subnetwork is extracted by choosing top eigen central nodes
(based on the supplied cut-off) and also the nodes adjacent to them.
}
\examples{
\dontrun{
data(KidneyGraphs)
g <- KidneyGraphs[[1]]
getSubnetworksEigenadjacent(g, eigenSortcutoff = 0.1)}
}
