% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionComp.R
\name{getTransitionmatrix}
\alias{getTransitionmatrix}
\title{Transition Matrix computation}
\usage{
getTransitionmatrix(g, walk = 1)
}
\arguments{
\item{g}{graph in igraph format}

\item{walk}{random walk distance (default=1)}
}
\value{
Transition matrix of specified walk
}
\description{
Transition Matrix computation
}
\examples{
\dontrun{
g <- igraph::make_tree(10, mode = "out")
# calculate one walk transition probability matrix
getTransitionmatrix(g, walk = 1)}
}
