% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGraphs.R
\docType{data}
\name{KidneyGraphs}
\alias{KidneyGraphs}
\title{Metabolic graphs of Kidney cancer samples from TCGA}
\format{
A `list` of  graphs in igraph format
}
\usage{
data(KidneyGraphs)
}
\description{
40 directed graphs with 441 nodes
}
\keyword{datasets}
