% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceComp.R
\name{getGraphlistdistance4Parts}
\alias{getGraphlistdistance4Parts}
\title{getGraphlistdistance4Parts}
\usage{
getGraphlistdistance4Parts(distribMatlist, index = 1, returnnodeDist = FALSE)
}
\arguments{
\item{distribMatlist}{list of probability distribution matrices}

\item{index}{possible values 1-4}

\item{returnnodeDist}{Return node-wise distances(default=FALSE), if TRUE returns
a list containing 2 elements: 1) nodewise distances and 2) graph distance
(Gram) matrix}
}
\value{
Gram matrix with Pairwise distances
}
\description{
used when there are numerous graphs, the graph samples can be
divided into four as specified by the index. Run in 4 separate R sessions.
Save the 4 matrices and reassemble the gram matrix
}
\examples{
\dontrun{
data("KidneyGraphs")
binsList <- getBins(KidneyGraphs)
nddList <- lapply(KidneyGraphs, function(x) getNodeDistanceDistr(x, binsList))
# Run the function in 4 separate terminals store the matrix and reassemble with
extractDist4part
nddDistanceMatPar1 <- getGraphlistdistance4Parts(nddList, index = 1)
# or use getGraphdistance4PartsParallel()}
}
\author{
Ichcha Manipur & Mario R Guarracino
}
