% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceComp.R
\name{getGraphpairdistance}
\alias{getGraphpairdistance}
\title{Find the Jensen Shannon divergence between distribution matrices of graphs}
\usage{
getGraphpairdistance(distrib1, distrib2)
}
\arguments{
\item{distrib1}{Probability distribution matrix of graph 1}

\item{distrib2}{Probability distribution matrix of graph 2}
}
\description{
Find the Jensen Shannon divergence between distribution matrices of graphs
}
\examples{
\dontrun{
g1 <- make_tree(10, mode = "out")
g2 <- make_tree(10, children = 3, mode = "out")
getGraphpairdistance(g1, g2)}
}
